Add files via upload
Download the CountryInfo_Student app from Moodle

Tasks:  

Display the list of currencies and languages for the selected country separate by a comma.  (3m)
Retrieve the image of the selected country flag from https://countryflagsapi.com and display it in an imageView that should be located in the Country Details activity. (4m)
Add a button to the country details UI with the text: WIKI + selected country name (e.g. WIKI MEXICO). Launch another Activity containing a WebView that displays the Wikipedia page for the selected country when this button is clicked. Include a home button back to the Country Details Activity.  (3m)